% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_bubble_co_Gnm.}
\alias{gg_bubble_co_Gnm.}
\title{Points inside Colombia's deptos map
Points inside Colombia's deptos map}
\usage{
gg_bubble_co_Gnm.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, prop_text = 0.5, color_point = "red",
  leg_pos = "right", text = FALSE, text_size = 2, color_map = "gray",
  color_frontier = "white", scale_point = 0.25, alpha = 0.5, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Points inside Colombia's deptos map
Points inside Colombia's deptos map
}
\section{ftypes}{
 Gnm
}

\examples{
add(1, 1)
add(10, 1)
}
