% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_bubble_depto_Gnm.}
\alias{gg_bubble_depto_Gnm.}
\title{Points inside Colombia's mpios map
Points inside Colombia's mpios map}
\usage{
gg_bubble_depto_Gnm.(data, titleLabel = "", subtitle = "", caption = "",
  depto_ = "05", leg_pos = "right", prop_text = 0.1, fillLabel = NULL,
  color_point = "red", text = FALSE, text_size = 2, color_map = "gray",
  color_frontier = "white", scale_point = 0.25, alpha = 0.5, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Points inside Colombia's mpios map
Points inside Colombia's mpios map
}
\section{ftypes}{
 Gnm
}

\examples{
add(1, 1)
add(10, 1)
}
