% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_bubble_GcdLonLat.}
\alias{gg_bubble_GcdLonLat.}
\title{Bubbles map
Bubbles map}
\usage{
gg_bubble_GcdLonLat.(data = NULL, mapName, opts = NULL, ...)
}
\arguments{
\item{x}{longitud.}

\item{y}{latitud.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubbles map
Bubbles map
}
\section{ftypes}{
 pLon-Lat
}

\examples{
NULL
}
