% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_bubble_latam_Gnm.}
\alias{gg_bubble_latam_Gnm.}
\title{Points inside Latam map
Points inside Latam map}
\usage{
gg_bubble_latam_Gnm.(data, titleLabel = "", subtitle = "", caption = "",
  fillLabel = NULL, color_point = "red", leg_pos = "right",
  color_map = "gray", color_frontier = "white", scale_point = 0.25,
  alpha = 0.5, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Points inside Latam map
Points inside Latam map
}
\section{ftypes}{
 Gnm
}

\examples{
add(1, 1)
add(10, 1)
}
