% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GcdNum.R
\name{gg_bubbles_Gcd}
\alias{gg_bubbles_Gcd}
\title{Ggplot bubbless by numerical variable}
\usage{
gg_bubbles_Gcd(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot bubbless by numerical variable
}
\section{ctypes}{
 Gcd
}

\examples{
gg_bubbles_Gcd(sample_data("Gcd", nrow = 10))
}
