% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{gg_bubbles_GlnGlt}
\alias{gg_bubbles_GlnGlt}
\alias{gg_bubbles_map_GlnGlt}
\title{Ggplot bubbles by latitud and longitud}
\usage{
gg_bubbles_map_GlnGlt(data = NULL, mapName = "world_countries",
  opts = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot bubbles by latitud and longitud
}
\section{ctypes}{
 Gln-Glt
}

\examples{
gg_bubbles_map_GlnGlt(sampleData("Gln-Glt", nrow = 10))
}
