% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GcdCatNum.R, R/bubbles_GnmCatNum.R,
%   R/bubbles_GnmNum.R
\name{gg_bubbles_Gnm}
\alias{gg_bubbles_Gnm}
\alias{gg_bubbles_GcdCat}
\alias{gg_bubbles_GnmCat}
\title{Ggplot bubbles by categorical variable}
\usage{
gg_bubbles_GcdCat(data = NULL, ...)

gg_bubbles_GnmCat(data = NULL, ...)

gg_bubbles_Gnm(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz

leaflet viz

leaflet viz
}
\description{
Ggplot bubbles by categorical variable

Ggplot bubbless by numerical variable

Ggplot bubbless by numerical variable
}
\section{ctypes}{
 Gcd-Cat

 Gnm

 Gnm
}

\examples{
gg_bubbles_GcdCat(sample_data("Gcd-Cat", nrow = 10))
gg_bubbles_GnmCat(sample_data("Gnm-Cat", nrow = 10))
gg_bubbles_Gnm(sample_data("Gnm", nrow = 10))
}
