% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GnmCatNum.R
\name{gg_bubbles_GnmCatNum}
\alias{gg_bubbles_GnmCatNum}
\title{Ggplot bubbless by categorical variable}
\usage{
gg_bubbles_GnmCatNum(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot bubbless by categorical variable
}
\section{ctypes}{
 Gnm-Cat-Num
}

\examples{
gg_bubbles_GnmCatNum(sample_data("Gnm-Cat-Num", nrow = 10))
}
