% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GnmNum.R
\name{gg_bubbles_GnmNum}
\alias{gg_bubbles_GnmNum}
\title{Ggplot bubbless by numerical variable}
\usage{
gg_bubbles_GnmNum(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot bubbless by numerical variable
}
\section{ctypes}{
 Gnm-Num
}

\examples{
gg_bubbles_GnmNum(sample_data("Gnm-Num", nrow = 10))
}
