% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{gg_bubbles_map_GlnGltCat.}
\alias{gg_bubbles_map_GlnGltCat.}
\title{Bubbles map}
\usage{
gg_bubbles_map_GlnGltCat.(data = NULL, mapName = "world_countries",
  title = NULL, subtitle = NULL, caption = NULL, minSize = 0.3,
  maxSize = 10, count = FALSE, legend = list(), border = list(),
  titleStyle = list(), labelWrap = 12, agg = "sum", fill = list(),
  marks = c(".", ","), nDigits = NULL, projections = list(),
  percentage = FALSE, format = c("", ""), theme = NULL)
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubbles map
}
\section{ftypes}{
 Gcd-Num
}

