% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bogota.R
\name{gg_choropleth_bogota_map_GcdNum.}
\alias{gg_choropleth_bogota_map_GcdNum.}
\alias{gg_choropleth_bogota_GcdNum.}
\title{Choropleth map Bogota's locations whitout Sumpaz
Choropleth Bogota's locations}
\usage{
gg_choropleth_bogota_GcdNum.(data = NULL, opts = list(titleLabel = "",
  subtitle = "", caption = "", reverse = FALSE, fillLabel = "", text = TRUE,
  text_size = 2, prop_text = "only_data", leg_pos = "right", titleLeg = "",
  color_map = "gray", color_frontier = "white"))
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth map Bogota's locations whitout Sumpaz
Choropleth Bogota's locations
}
\section{ftypes}{
 Gcd-Num
}

\examples{
NULL
}
