% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_choropleth_co_GnmNum.}
\alias{gg_choropleth_co_GnmNum.}
\title{Choropleth of Colombia's deptos
Choropleth of Colombia's deptos}
\usage{
gg_choropleth_co_GnmNum.(data, titleLabel = "", subtitle = "",
  caption = "", reverse = FALSE, text_size = 2, fillLabel = NULL,
  leg_pos = "right", text = FALSE, prop_text = 0.5, color_map = "gray",
  color_frontier = "white", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth of Colombia's deptos
Choropleth of Colombia's deptos
}
\section{ftypes}{
 Gnm-Num
}

\examples{
add(1, 1)
add(10, 1)
}
