% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_choropleth_depto_GnmNum.}
\alias{gg_choropleth_depto_GnmNum.}
\title{Choropleth by filtering deptos
Choropleth by filtering deptos}
\usage{
gg_choropleth_depto_GnmNum.(data, titleLabel = "", subtitle = "",
  caption = "", depto_ = "05", reverse = FALSE, prop_text = 0.1,
  fillLabel = NULL, leg_pos = "right", text = FALSE, text_size = 2,
  color_map = "gray", color_frontier = "white", ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth by filtering deptos
Choropleth by filtering deptos
}
\section{ftypes}{
 Gnm-Num
}

\examples{
add(1, 1)
add(10, 1)
}
