% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{gg_choropleth_Gcd}
\alias{gg_choropleth_Gcd}
\alias{gg_choropleth_Gcd.}
\title{Ggplot choropleths by geographical code}
\usage{
gg_choropleth_Gcd.(data = NULL, mapName = "world_countries",
  title = NULL, subtitle = NULL, caption = NULL, count = TRUE,
  legend = list(), border = list(), labelWrap = 12, agg = "sum",
  fill = list(), marks = c(".", ","), nDigits = NULL,
  projections = list(), percentage = FALSE, format = c("", ""),
  theme = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot choropleths by geographical code
}
\section{ctypes}{
 Gcd
}

\examples{
gg_choropleth_Gcd(sampleData("Gcd", nrow = 10))
}
