% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GcdCatNum.R
\name{gg_choropleth_GcdCatNum}
\alias{gg_choropleth_GcdCatNum}
\title{Ggplot choropleths by categorical variable}
\usage{
gg_choropleth_GcdCatNum(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot choropleths by categorical variable
}
\section{ctypes}{
 Gcd-Cat-Num
}

\examples{
gg_choropleth_GcdCatNum(sample_data("Gcd-Cat-Num", nrow = 1000))
}
