% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GnmNum.R
\name{gg_choropleth_Gnm}
\alias{gg_choropleth_Gnm}
\title{Ggplot choropleths by numerical variable}
\usage{
gg_choropleth_Gnm(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot choropleths by numerical variable
}
\section{ctypes}{
 Gnm
}

\examples{
gg_choropleth_Gnm(sample_data("Gnm", nrow = 10))
}
