% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GnmCatNum.R
\name{gg_choropleth_GnmCat}
\alias{gg_choropleth_GnmCat}
\title{Ggplot choropleths by categorical variable}
\usage{
gg_choropleth_GnmCat(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot choropleths by categorical variable
}
\section{ctypes}{
 Gnm-Cat
}

\examples{
gg_choropleth_GnmCat(sample_data("Gnm-Cat", nrow = 10))
}
