% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GnmCatNum.R
\name{gg_choropleth_GnmCatNum}
\alias{gg_choropleth_GnmCatNum}
\title{Ggplot choropleths by categorical variable}
\usage{
gg_choropleth_GnmCatNum(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot choropleths by categorical variable
}
\section{ctypes}{
 Gnm-Cat-Num
}

\examples{
gg_choropleth_GnmCatNum(sample_data("Gnm-Cat-Num", nrow = 1000))
}
