% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GnmNum.R
\name{gg_choropleth_GnmNum}
\alias{gg_choropleth_GnmNum}
\title{Ggplot choropleths by numerical variable}
\usage{
gg_choropleth_GnmNum(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot choropleths by numerical variable
}
\section{ctypes}{
 Gnm-Num
}

\examples{
gg_choropleth_GnmNum(sampleData("Gnm-Num", nrow = 10))
}
