% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_choropleth_latam_GnmNum.}
\alias{gg_choropleth_latam_GnmNum.}
\title{Choropleth of Latam
Choropleth of Latam}
\usage{
gg_choropleth_latam_GnmNum.(data, titleLabel = "", subtitle = "",
  caption = "", reverse = FALSE, fillLabel = NULL, leg_pos = "right",
  text = TRUE, prop_text = "only_data", color_map = "gray",
  color_frontier = "white", text_size = 2, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth of Latam
Choropleth of Latam
}
\section{ftypes}{
 Gnm-Num
}

\examples{
add(1, 1)
add(10, 1)
}
