% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{gg_choropleth_map_GcdCat.}
\alias{gg_choropleth_map_GcdCat.}
\title{Choropleth map
Choropleth map}
\usage{
gg_choropleth_map_GcdCat.(data = NULL, mapName = "world_countries",
  title = NULL, subtitle = NULL, caption = NULL, count = TRUE,
  legend = list(), border = list(), labelWrap = 12, fill = list(),
  marks = c(".", ","), projections = list(), theme = NULL)
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth map
Choropleth map
}
\section{ftypes}{
 Gcd-Cat
}

