% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{gg_choropleth_map_GcdNum.}
\alias{gg_choropleth_map_GcdNum.}
\title{Choropleth map
Choropleth map}
\usage{
gg_choropleth_map_GcdNum.(data = NULL, mapName, opts = NULL, ...)
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth map
Choropleth map
}
\section{ftypes}{
 Gcd-Num
}

\examples{
NULL
}
