% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{gg_choropleth_map_GnmCat.}
\alias{gg_choropleth_map_GnmCat.}
\title{Choropleth map
Choropleth map}
\usage{
gg_choropleth_map_GnmCat.(data = NULL, mapName = "world_countries",
  title = NULL, subtitle = NULL, caption = NULL, count = FALSE,
  legend = list(), border = list(), titleStyle = list(),
  labelWrap = 12, agg = "sum", fill = list(), marks = c(".", ","),
  nDigits = NULL, projections = list(), percentage = FALSE,
  format = c("", ""), theme = NULL)
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth map
Choropleth map
}
\section{ftypes}{
 Gnm-Cat
}

