% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_choropleth_map_GnmNum.}
\alias{gg_choropleth_map_GnmNum.}
\title{Choropleth map
Choropleth map}
\usage{
gg_choropleth_map_GnmNum.(data = NULL, mapName, opts = list(titleLabel = "",
  subtitle = "", caption = "", reverse = FALSE, fillLabel = NULL, text = TRUE,
  text_size = 2, prop_text = "only_data", leg_pos = "right", titleLeg = "",
  color_map = "gray", color_frontier = "white"))
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth map
Choropleth map
}
\section{ftypes}{
 Gnm-Num
}

