% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_sketchmap_depto_GnmNum.}
\alias{gg_sketchmap_depto_GnmNum.}
\title{Sketch map departments
Sketch map departments}
\usage{
gg_sketchmap_depto_GnmNum.(depto_ = depto_, color_map = "gold",
  color_frontier = "black")
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Sketch map departments
Sketch map departments
}
\section{ftypes}{
 Gnm-Num
}

\examples{
add(1, 1)
add(10, 1)
}
