% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_bubble_latam_Gnm.}
\alias{gg_bubble_latam_Gnm.}
\alias{gg_bubble_depto_Gnm.}
\title{Points inside Latam map
Points inside Latam map}
\usage{
gg_bubble_depto_Gnm.(data, titleLabel = "", subtitle = "", caption = "",
  depto_ = "05", leg_pos = "right", prop_text = 0.1, fillLabel = NULL,
  color_point = "red", text = FALSE, text_size = 2, color_map = "gray",
  color_frontier = "white", scale_point = 0.25, alpha = 0.5, ...)
}
\arguments{
\item{x}{A category.}

\item{y}{A number.}

\item{x}{A category.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
Points inside Latam map
Points inside Latam map
}
\section{ftypes}{
 Gnm

 Gnm
}

\examples{
add(1, 1)
#' add(10, 1)
gg_bubble_latam_Gnm. <- function(data, titleLabel = "", subtitle = "", caption = "", fillLabel = NULL,
                                 color_point = "red", leg_pos = "right",
                                 color_map = "gray", color_frontier = "white", scale_point = 0.25,
                                 alpha = 0.5, ...){

  f <- fringe(data)
  nms <- getClabels(f)
  flab <- fillLabel \%||\% nms[3]
  data <- f$d

  options(warn=-1)
  data_latam <- suppressMessages(read_csv(system.file("geodata/latam/latam.csv", package = "geodata"), col_names = TRUE))

  data_graph <- data \%>\% dplyr::group_by(a, b) \%>\% dplyr::summarise(count = n())

  graph <- ggplot(data_latam) +
    geom_map(map = data_latam,
             aes(map_id = id, x = long, y = lat, group = group), fill = color_map,
             color=color_frontier, size = 0.25) +
    expand_limits(x = data_latam$long, y = data_latam$lat) +
    coord_fixed()

  graph <- graph + geom_point(data = data_graph, aes(x = a, y = b),
                              size = data_graph$count * scale_point,
                              colour = color_point, alpha = alpha) + coord_map() + coord_fixed()  +
    labs(x = "", y = "", title = titleLabel, subtitle = subtitle, caption = caption) + theme_ds() + theme_ds_clean() +
    theme(legend.position=leg_pos)
  options(warn=0)

  graph
}
Points inside Colombia's mpios map
Points inside Colombia's mpios map
add(1, 1)
add(10, 1)
}
