% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{gg_bubbles_GcdNum}
\alias{gg_bubbles_GcdNum}
\title{Bubbles map
Bubbles map}
\usage{
gg_bubbles_GcdNum(data = NULL, mapName = "world_countries", opts = NULL)
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubbles map
Bubbles map
}
\section{ftypes}{
 Gcd-Num
}

\examples{
NULL
}
