% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GlnGltNum.R
\name{gg_bubbles_GlnGlt}
\alias{gg_bubbles_GlnGlt}
\title{Ggplot2 bubbles by numerical variable}
\usage{
gg_bubbles_GlnGlt(data = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot2 bubbles by numerical variable
}
\section{ctypes}{
 Gln-Glt
}

\examples{
gg_bubbles_GlnGlt(sample_data("Gln-Glt", nrow = 10))
}
