% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{gg_bubbles_GlnGltNum}
\alias{gg_bubbles_GlnGltNum}
\alias{gg_bubbles_map_GlnGltNum.}
\title{Ggplot bubbles by latitud and longitud}
\usage{
gg_bubbles_map_GlnGltNum.(data = NULL, mapName = "world_countries",
  title = NULL, subtitle = NULL, caption = NULL, minSize = 0.1,
  maxSize = 10, legend = list(), border = list(),
  titleStyle = list(), labelWrap = 12, agg = "sum", fill = list(),
  marks = c(".", ","), nDigits = NULL, projections = list(),
  percentage = FALSE, format = c("", ""), theme = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot bubbles by latitud and longitud
}
\section{ctypes}{
 Gln-Glt-Num
}

\examples{
gg_bubbles_map_GlnGltNum.(sampleData("Gln-Glt-Num", nrow = 10))
}
