% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{gg_bubbles_GlnGltNum}
\alias{gg_bubbles_GlnGltNum}
\title{Ggplot bubbles by latitud and longitud}
\usage{
gg_bubbles_GlnGltNum(data = NULL, mapName = "world_countries", opts = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot bubbles by latitud and longitud
}
\section{ctypes}{
 Gln-Glt-Num
}

\examples{
gg_bubbles_GlnGltNum(sampleData("Gln-Glt-Num", nrow = 10))
}
