% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_GnmNum.R
\name{gg_bubbles_Gnm}
\alias{gg_bubbles_Gnm}
\title{Ggplot bubbless by numerical variable}
\usage{
gg_bubbles_Gnm(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot bubbless by numerical variable
}
\section{ctypes}{
 Gnm
}

\examples{
gg_bubbles_Gnm(sample_data("Gnm", nrow = 10))
}
