% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{gg_bubbles_GnmNum}
\alias{gg_bubbles_GnmNum}
\title{Bubbles map}
\usage{
gg_bubbles_GnmNum(data = NULL, mapName = "world_countries", opts = NULL)
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubbles map
}
\section{ftypes}{
 Gnm-Num
}

