% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{gg_bubbles_map_GnmNum.}
\alias{gg_bubbles_map_GnmNum.}
\title{Bubbles map
Bubbles map}
\usage{
gg_bubbles_map_GnmNum.(data = NULL, mapName, opts = list(titleLabel = "",
  subtitle = "", caption = "", fillLabel = NULL, prop_text = "only_data",
  color_point = "red", leg_pos = "right", titleLeg = "", text = FALSE, text_size
  = 2, color_map = "gray", color_frontier = "white", scale_point = 3, alpha =
  0.5))
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Bubbles map
Bubbles map
}
\section{ftypes}{
 Gnm-Num
}

