% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GcdNum.R
\name{gg_choropleth_Gcd}
\alias{gg_choropleth_Gcd}
\title{Ggplot choropleths by numerical variable}
\usage{
gg_choropleth_Gcd(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot choropleths by numerical variable
}
\section{ctypes}{
 Gcd
}

\examples{
gg_choropleth_Gcd(sample_data("Gcd", nrow = 10))
}
