% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GcdCatNum.R
\name{gg_choropleth_GcdCat}
\alias{gg_choropleth_GcdCat}
\title{Ggplot choropleths by categorical variable}
\usage{
gg_choropleth_GcdCat(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot choropleths by categorical variable
}
\section{ctypes}{
 Gcd-Cat
}

\examples{
gg_choropleth_GcdCat(sample_data("Gcd-Cat", nrow = 10))
}
