% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{gg_choropleth_GcdNum}
\alias{gg_choropleth_GcdNum}
\title{Choropleth map
Choropleth map}
\usage{
gg_choropleth_GcdNum(data = NULL, mapName = "world_countries", opts = NULL)
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth map
Choropleth map
}
\section{ftypes}{
 Gcd-Num
}

\examples{
gg_choropleth_GcdNum()
gg_choropleth_GcdNum(data = sampleData("Gcd-Num"))
gg_choropleth_GcdNum(data = sampleData("Gcd-Num", 1000), mapName = "american_countries")
}
