% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{gg_choropleth_Gnm}
\alias{gg_choropleth_Gnm}
\title{Ggplot choropleths by geographical code}
\usage{
gg_choropleth_Gnm(data = NULL, mapName = "world_countries", opts = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
leaflet viz
}
\description{
Ggplot choropleths by geographical code
}
\section{ctypes}{
 Gnm
}

\examples{
gg_choropleth_Gnm(sampleData("Gnm", nrow = 10))
}
