% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{gg_choropleth_GnmCat}
\alias{gg_choropleth_GnmCat}
\title{Choropleth map
Choropleth map}
\usage{
gg_choropleth_GnmCat(data = NULL, mapName = "world_countries",
  opts = NULL)
}
\arguments{
\item{x}{A code.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Choropleth map
Choropleth map
}
\section{ftypes}{
 Gnm-Cat
}

